KTL-1731: revert to default Spring Boot validator, validate compose-wasm imperatively
