+++ common/build.gradle.kts
@@ -5,5 +5,4 @@ plugins {
 dependencies {
     implementation(libs.kotlin.compiler.embeddable)
     implementation(libs.jackson.module.kotlin)
-    implementation(libs.jakarta.validation.api)
 }
