+++ src/main/kotlin/com/compiler/server/validation/AbstractCompilerArgumentsValidator.kt
@@ -5,7 +5,6 @@ import com.compiler.server.model.ExtendedCompilerArgument
 import com.compiler.server.model.ExtendedCompilerArgumentValue
 import com.compiler.server.model.ListExtendedCompilerArgumentValue
 import com.compiler.server.model.StringExtendedCompilerArgumentValue
-import jakarta.validation.ConstraintValidatorContext
 
 /**
  * @param knownCompilerArguments The set of predefined compiler arguments that are
